requirements.txt
praw
python-dotenv
pymysql
